requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
PyJWT
datetime
requests
aiohttp
asyncio
cryptography>=3.3.1
skyflow
gunicorn
